More intuitive package restructuring and decoupling.
Better package organization, and the translation abilities of
QueryRequest and QueryResult factored out into Translator classes.
