+++ src/main/java/echoquery/SingletonConnections.java
@@ -1,4 +1,4 @@
-package echoquery.sql;
+package echoquery;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
