+++ src/main/java/echoquery/querier/QueryBuildException.java
@@ -1,4 +1,4 @@
-package echoquery.sql;
+package echoquery.querier;
 
 public class QueryBuildException extends Exception {
