+++ src/main/java/echoquery/querier/ResultTable.java
@@ -1,4 +1,4 @@
-package echoquery.sql;
+package echoquery.querier;
 
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
