+++ src/main/java/echoquery/querier/schema/ColumnType.java
@@ -1,4 +1,4 @@
-package echoquery.sql.model;
+package echoquery.querier.schema;
 
 public enum ColumnType {
   NUMBER,
