+++ src/main/java/echoquery/querier/schema/ColumnValue.java
@@ -1,4 +1,4 @@
-package echoquery.sql.model;
+package echoquery.querier.schema;
 
 public class ColumnValue {
   private String value;
